requirements.txt
transformers==4.41.2
torch==2.2.2
langchain
google-generativeai
python-dotenv
requests
streamlit
pillow
